pyproject.toml
[project]
name = "temporalemailmanager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp==3.9.5",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "async-timeout==4.0.3",
    "attrs==23.2.0",
    "blinker==1.8.2",
    "click==8.1.7",
    "exceptiongroup==1.3.0",
    "flask==3.0.3",
    "frozenlist==1.4.1",
    "idna==3.7",
    "iniconfig==2.0.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.4",
    "markupsafe==2.1.5",
    "multidict==6.0.5",
    "packaging==24.0",
    "pluggy==1.5.0",
    "protobuf==5.27.0",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pytest==8.2.2",
    "pytest-asyncio==0.23.7",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "ruff>=0.12.0",
    "six==1.17.0",
    "temporalio==1.10.0",
    "tomli==2.2.1",
    "types-protobuf==5.26.0.20240422",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "werkzeug==3.0.3",
    "yarl==1.9.4",
]
